requirements.txt
allure-pytest==2.9.45
pytest==7.1.2
requests==2.27.1